Move last src integration tests to tests and fix 5.2 php lint
