--- integration-tests/src/integration-tests/config/database-migration-test.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Yoast\Tests\IntegrationTests\Config;
-
-use Yoast\WP\Free\Config\Database_Migration;
-
-/**
- * Class Database_Migration_Test.
- *
- * @group   db-migrations
- *
- * @package Yoast\Tests
- */
-class Database_Migration_Test extends \PHPUnit_Framework_TestCase {
-
-
-}
