+++ tests/doubles/indexable-author-formatter-double.php
