+++ tests/doubles/indexable-post-formatter-double.php
