+++ tests/doubles/indexable-term-formatter-double.php
