+++ tests/doubles/oauth/client.php
