+++ tests/doubles/plugin.php
